requirements.txt
# ── Desktop App (Current Version) ────────────────────────────────────────────
mysql-connector-python==8.3.0
bcrypt==4.1.3
python-dotenv==1.0.1

# ── FastAPI Backend (Week 1 Upgrade — install when starting backend) ──────────
# fastapi==0.111.0
# uvicorn[standard]==0.29.0
# sqlalchemy==2.0.30
# alembic==1.13.1
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4
# python-multipart==0.0.9
# pydantic==2.7.1
# pydantic-settings==2.2.1
# reportlab==4.1.0
# pytest==8.2.0
# httpx==0.27.0
